KSCurrency refactor (#528)
* Using doubles instead of floats for currency.

* I added a Model for `Country` based off https://github.com/kickstarter/ios-oss/blob/master/KsApi/models/Project.Country.swift.

* We round down for project pledged and goal amounts. But round half_up (normally) when viewing backing or converted amounts.

* updated reward tests to use possible minimums and show new rounding

* the default rounding is DOWN so we shouldn't ever pass that in

+++ app/src/main/java/com/kickstarter/models/Backing.java
@@ -16,7 +16,7 @@ import auto.parcel.AutoParcel;
 @AutoGson
 @AutoParcel
 public abstract class Backing implements Parcelable {
-  public abstract float amount();
+  public abstract double amount();
   public abstract @Nullable User backer();
   public abstract @Nullable String backerNote();
   public abstract long backerId();
@@ -36,7 +36,7 @@ public abstract class Backing implements Parcelable {
 
   @AutoParcel.Builder
   public abstract static class Builder {
-    public abstract Builder amount(float __);
+    public abstract Builder amount(double __);
     public abstract Builder backer(User __);
     public abstract Builder backerNote(String __);
     public abstract Builder backerId(long __);
